SDK-537: Move dynamic_sharing_service into subpackage
Move errors.go and validate.go into subpackage to allow use between
other subpackages
