+++ dynamic_sharing_service/dynamic_scenario_builder.go
@@ -1,4 +1,4 @@
-package yoti
+package dynamic_sharing_service
 
 import (
 	"encoding/json"
