+++ dynamic_sharing_service/source_constraint_builder_test.go
@@ -1,4 +1,4 @@
-package yoti
+package dynamic_sharing_service
 
 import (
 	"fmt"
