+++ dynamic_sharing_service/wanted_anchor_builder.go
@@ -1,4 +1,4 @@
-package yoti
+package dynamic_sharing_service
 
 import (
 	"encoding/json"
